+++ features/ajna/positions/earn/components/AjnaEarnSlider.tsx
@@ -1,128 +1,17 @@
 import BigNumber from 'bignumber.js'
-import { WAD_PRECISION } from 'components/constants'
 import { SliderValuePicker } from 'components/dumb/SliderValuePicker'
 import { PillAccordion } from 'components/PillAccordion'
-import {
-  ajnaDefaultMarketPriceOffset,
-  ajnaDefaultPoolRangeMarketPriceOffset,
-} from 'features/ajna/common/consts'
 import { useAjnaGeneralContext } from 'features/ajna/positions/common/contexts/AjnaGeneralContext'
 import { useAjnaProductContext } from 'features/ajna/positions/common/contexts/AjnaProductContext'
 import { AjnaEarnInput } from 'features/ajna/positions/earn/components/AjnaEarnInput'
 import { AJNA_LUP_MOMP_OFFSET } from 'features/ajna/positions/earn/consts'
+import { convertSliderThresholds } from 'features/ajna/positions/earn/helpers/convertSliderThresholds'
+import { getMinMaxAndRange } from 'features/ajna/positions/earn/helpers/getMinMaxAndRange'
+import { snapToPredefinedValues } from 'features/ajna/positions/earn/helpers/snapToPredefinedValues'
 import { formatAmount, formatDecimalAsPercent } from 'helpers/formatters/format'
-import { one, zero } from 'helpers/zero'
+import { one } from 'helpers/zero'
 import { useTranslation } from 'next-i18next'
 import React, { useEffect, useMemo } from 'react'
-
-function snapToPredefinedValues(value: BigNumber, predefinedSteps: BigNumber[]) {
-  return predefinedSteps.reduce((prev, curr) => {
-    return curr.minus(value).abs().lt(prev.minus(value).abs()) ? curr : prev
-  })
-}
-
-function getMinMaxAndRange({
-  highestThresholdPrice,
-  lowestUtilizedPrice,
-  lowestUtilizedPriceIndex,
-  mostOptimisticMatchingPrice,
-  marketPrice,
-  offset, // 0 - 1, percentage value
-}: {
-  highestThresholdPrice: BigNumber
-  lowestUtilizedPrice: BigNumber
-  lowestUtilizedPriceIndex: BigNumber
-  mostOptimisticMatchingPrice: BigNumber
-  marketPrice: BigNumber
-  offset: number
-}) {
-  // check whether pool contain liquidity and borrowers, if no generate default range from the lowest price to market price
-  if (lowestUtilizedPriceIndex.eq(zero)) {
-    const defaultRange = [marketPrice.times(one.minus(ajnaDefaultPoolRangeMarketPriceOffset))]
-
-    while (
-      defaultRange[defaultRange.length - 1].lt(
-        marketPrice.times(one.minus(ajnaDefaultMarketPriceOffset)),
-      )
-    ) {
-      defaultRange.push(
-        defaultRange[defaultRange.length - 1].times(1.005).decimalPlaces(WAD_PRECISION),
-      )
-    }
-
-    return {
-      min: defaultRange[0],
-      max: defaultRange[defaultRange.length - 1],
-      range: defaultRange,
-    }
-  }
-
-  // Generate ranges from min to lup
-  const lupNearHtpRange = [lowestUtilizedPrice]
-
-  while (lupNearHtpRange[lupNearHtpRange.length - 1].gt(highestThresholdPrice)) {
-    lupNearHtpRange.push(lupNearHtpRange[lupNearHtpRange.length - 1].div(1.005))
-  }
-
-  const nearHtpMinRange = [lupNearHtpRange[lupNearHtpRange.length - 1]]
-
-  while (
-    nearHtpMinRange[nearHtpMinRange.length - 1].gt(nearHtpMinRange[0].times(one.minus(offset)))
-  ) {
-    nearHtpMinRange.push(nearHtpMinRange[nearHtpMinRange.length - 1].div(1.005))
-  }
-
-  // Generate ranges from lup to max
-  const lupNearMompRange = [lowestUtilizedPrice]
-
-  while (lupNearMompRange[lupNearMompRange.length - 1].lt(mostOptimisticMatchingPrice)) {
-    lupNearMompRange.push(lupNearMompRange[lupNearMompRange.length - 1].times(1.005))
-  }
-
-  const nearMompMaxRange = [lupNearMompRange[lupNearMompRange.length - 1]]
-
-  while (
-    nearMompMaxRange[nearMompMaxRange.length - 1].lt(nearMompMaxRange[0].times(one.plus(offset)))
-  ) {
-    nearMompMaxRange.push(nearMompMaxRange[nearMompMaxRange.length - 1].times(1.005))
-  }
-
-  const range = [
-    ...new Set([...nearHtpMinRange, ...lupNearHtpRange, ...lupNearMompRange, ...nearMompMaxRange]),
-  ]
-    .sort((a, b) => a.toNumber() - b.toNumber())
-    .map((item) => item.decimalPlaces(18))
-
-  return {
-    min: range[0],
-    max: range[range.length - 1],
-    range,
-  }
-}
-
-function convertSliderValuesToPercents(value: BigNumber, min: BigNumber, max: BigNumber) {
-  return value.minus(min).div(max.minus(min)).times(100).toNumber()
-}
-
-function convertSliderThresholds({
-  min,
-  max,
-  highestThresholdPrice,
-  lowestUtilizedPrice,
-  mostOptimisticMatchingPrice,
-}: {
-  min: BigNumber
-  max: BigNumber
-  highestThresholdPrice: BigNumber
-  lowestUtilizedPrice: BigNumber
-  mostOptimisticMatchingPrice: BigNumber
-}) {
-  return {
-    htpPercentage: convertSliderValuesToPercents(highestThresholdPrice, min, max),
-    lupPercentage: convertSliderValuesToPercents(lowestUtilizedPrice, min, max),
-    mompPercentage: convertSliderValuesToPercents(mostOptimisticMatchingPrice, min, max),
-  }
-}
 
 export function AjnaEarnSlider({ isDisabled }: { isDisabled?: boolean }) {
   const { t } = useTranslation()
