+++ features/ajna/positions/earn/helpers/convertSliderValuesToPercents.ts
@@ -1,0 +1,5 @@
+import BigNumber from 'bignumber.js'
+
+export const convertSliderValuesToPercents = (value: BigNumber, min: BigNumber, max: BigNumber) => {
+  return value.minus(min).div(max.minus(min)).times(100).toNumber()
+}
